Create new game on db when clicking new game link
